ims_usrloc_pcscf:   implementation of db_mode DB_ONLY (value 3)
In order to support a redundant PCSCF configuration - i.e. a logical PCSCF
consists of 2 physical nodes (node1 and node2) some enhancements were
introduced for handling of contacts. Redundancy means that SIP messages
for a client are normally handled by node1 but in case node1 is not
reachable SIP messages are redirected to node2. Of course the DB_ONLY
mode must work also for single PCSCF node configuration. Important aspects
of this implementation are database integrity - i.e. avoid invalid table
entries (for example data which are expired long time ago or have
invalid states) - and keeping PCSCF cache in sync with database tables.
A wrapper was built for method get_pcontact which tries to find the pcontact
in cache and if search is not successful tries to download and insert from
db location table - also some effort is added here to find the pcontact if it exists in cache.
The contact expiry handler was modified to sync contact
expiry in cache with db location entry and in case of real contact expiry
sends PUBLISH to SCSCF to let NOTIFY finally delete the contact.
An audit for older expired pcontacts was introduced which cares for
getting rid of these contacts.
Some code was introduced to help registering callbacks for contacts
which are inserted into cache when being downloaded from database - for
example ims_qos callback as at the time of insertion the message that triggered
the original callback registering is long gone.