lost: new features LoST redirect and NAPTR lookup (LIS, LoST)
The extension of the module allows dynamic querying of LIS/HELD
and LOST services via NAPTR lookup. In the case of LOST, a redirect
response is evaluated. In case a lost_held_request (used to connect
to a LIS) is passed with an empty string ("") for the connection
parameter, then P-A-I or From header value hostnames are used for
NAPTR lookup to get a corresponding service.