core: fix to xavp_rm_internal (#2604)
This fixes the following issue:
https://github.com/kamailio/kamailio/issues/2604

Description of the issue:

When called to remove a specific index from a given xavp, function xavp_rm_by_index removes the index (as expected) but also all others before it.

E.g :
If called with idx = 1, it removes indexes 0 and 1.
Likewise if invoked with idx = 2 => the first 3 elements are removed.

This bug is located in function xavp_rm_internal. An assignment was missing when looping over the xavp list.
Same for xavi_rm_internal.

(cherry picked from commit 12414972ad0c28ac50ece3c14f98134c3f06c522)