Expose the domain in certificate validation errors
We ran into an issue recently with certificate validation and this modification to the source code helped us quickly identify the root cause.  Previously we had no way of seeing which domain was invalid.  I didn't like using malloc here but hopefully freeing it in the same context ensures no memory leak as well avoid any memory corruption... with the memcpy and '\0'... hope this helps...