tm: API improvement, new function introduced: t_append_branch_by_contact()
New module functions introduced:
- t_append_branch_by_contact()

This commit introduces a possibility of TM's API to append a branch
based on specific location (Contact's URI).

Indeed the 't_append_branch_by_contact()' is a corrected copy of
the 't_append_branches()', which instead takes Contact as a parameter
of str type.

The separate function has been introduced, in order to save full
back-compatibility, and not affect existing implementation.