dispatcher: Fix handling of inactive destination for alg 13
Alg 13 did try to distribute calls to inactive destinations.
If the highest priority destination is inactive, hash is not updated
and the xavp is not set. This is resulting in failover mechanism
not working at all for the given call.

When the hash variable is not updated, it makes alg 13 behave like
round robin if the scenario above occurs. If you got two destinations
and the highest priority is out of service, 50% of the calls will fail.

Now I tried a more simple approach updating hash with the first
entry of the sorted list.

(cherry picked from commit adba3caa2d12d1f8b89d720fbb9a6a9cdc8b2573)