dialog: Refactored dlg_get_matches and dlg.list_match RPC command to use a shared dlg_list_matches function.
This shared function will allow identical matching logic on both, so that any future enhancements apply to both features.
The dlg_list_matches function returns an array of dlg_cell_match, which contains the pointer to the original dialog, as
well as the table entry index. This is done because, since we are using the original dialog, locking must still be done
on the matches returned.