http_client: Add parameter connect_timeout_ms / timeout_ms (2)
The new parameter connect_timeout_ms (global) / timeout_ms (in httpcon) allows to specify a timeout in milliseconds on curl requests.
If this parameter is defined (non zero), then the timeout in seconds is ignored.

If either timeout or timeout_ms is defined at connection level, then they take precedence over the global parameters.
These principles also apply to the file configuration.

The timeout in ms is used internally to set CURLOPT_TIMEOUT_MS.
The value is shown in ms when using RPC "httpclient.listcon".