kamcmd: don't clash with ENV NAME or ctl.so module
NAME variable might be set to the current HOSTNAME in some shell and
also clash with the value set by the module calling MOD_INSTALL_UTILS by
passing a NAME variable.

With commit 1774cee62098 ("kamcmd: allow defining the name of the
application from command line") this resulted in the kamcmd bin being
renamed to all kind of name from hostname to ctl.so.

Fix this by checking the variable to a more safe name that is not
already defined in shell by default and also that doesn't clash with
module variables, use UTIL_NAME as an alternative to NAME.

UTIL_NAME now needs to be used to create kamcmd with custom name.

Fixes: 1774cee62098 ("kamcmd: allow defining the name of the application from command line")
Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>