core: dprint - include time in stderr json log
This patch introduces a new "time" field for JSON logging to stderr. The
CEE format already includes this field, but the Kamailio-specific format
doesn't. If you need timestamps in your structured logs, but don't want
to use the CEE format due to its limitations, this patch'd give you the
best of both worlds.