usrloc: check on db delete the return value of memchr
When inserting into the database the AOR is split at the @ sign and
if there is no @ sign in the AOR only the domain part is filled and the
user part is left empty.

But for deleting this is not done and the query failed to be executed
and the AOR is not deleted. This PR add this behaviour of only comparing against the
domain part if the AOR doesn't contain a @ sign.

(cherry picked from commit 49276a1f43b1a3bf4a8d681888df964000360c79)
(cherry picked from commit 28bbcd5a50f45b8217125481706180a76323dc44)