packaging: RPM build option to build just the geoip2 module
The geoip sub-package currently builds both geoip and geoip2 module. The former is based on the old MaxMind GeoIP library whose support officially ended in 2022 and is no longer available for new distributions, and specifically not in EPEL 9.

The new build option `with geoip2` builds the `geoip` sub-package with just the geoip2 module.