packaging: disable geoip module when it isn't buildable
Added an RPM build option to disable building with the obsolete MaxMind GeoIP library, when building geoip capabilities,
and use it by default on distributions that only ship the new MaxMind library - to only build the geoip2 module.

Due to the old MaxMind GeoIP library's support officially ending in 2022, it is no longer available for new distributions,
and specifically not in EPEL 9. The new build option `without geoip1` builds the `geoip` sub-package with just the geoip2 module.