kafka: Add interceptor for broker state
So far there is no message logged by librdkafka or kamailio when broker is
back up again. The only way to find out that broker is back up is to
constatntly check kamcmd stats when errors stop growing.

Broker state interceptor is available since librdkafka v2.0.0-RC1
(Dec 21 2022). So only if you are using that librdkafka version,
or newer than that, this code will be enabled.