db_redis: add replicas failover to master in sentinel and use_replicas scenarios
- in sentinel scenarios with master/replica setup, if use_replicas
is enabled and for whatever reason no replicas are available,
the module will fall back to connecting to the master (this behavior cannot
be toggled off)
- introduced recheck_replicas_interval module param: it sets the interval
in seconds to recheck for available replicas so that we don't remain
connected to the master. Because of implementation details, only the first
check can happen after half of this interval. Value 0 means no recheck will
be done.