pv: Always convert int to string in {s.int} transformation
PV's may hold a string as well as an integer, and the s.int transformation
always assumes that an existing string value always represents the integer.
This is not the case for some PV's such as $rP where the string value is
not equsal to the integer (ie, string "tls" and integer 3).

By ignoring the existing string value and always creating a new string from
the integer value, it is guaranteed that s.int always returns an numerical
value.

Fixes #4444